Good enough for now

+++ Assets/Code/Scripts/TargetPractice.cs
@@ -5,7 +5,7 @@ public class TargetPractice : MonoBehaviour, IDamagable
     [SerializeField] private Renderer rend;
     [SerializeField] private Material hitMat;
 
-    public void TakeDamage(float _damage)
+    public void TakeDamage(int _damage)
     {
         rend.material = hitMat;
         //Debug.Log($"{name} (parent: {transform.parent.name}) was hit!");
